Data
Overall data that is used in the analysis. It is explained in detail in the paper. Some of the variables are coming from textual and complex datasets and coded by hand here. It is combined 2 years ago. For further discussion and detail description of the dataset, please send an email.